+++ README.md
@@ -14,21 +14,9 @@ pnpm dev
 bun dev
 ```
 
-
-To sync your models run:
-```bash
-npm run sync-db
-```
-To seed your database with dates run:
-npm run seed-db
-```
-To seed:undo:all your database:
-npm run unseed-db
-```
-
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
-You can start editing the page by modifying `app/page.js`. The page auto-updates as you edit the file.
+You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
 
 This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
