--- next14ts/src/app/layout.tsx
@@ -1,19 +1,0 @@
-// import { Inter } from 'next/font/google'
-// import './globals.css'
-
-// const inter = Inter({ subsets: ['latin'] })
-
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="en">
-      <body
-          // className={inter.className}
-      >{children}</body>
-    </html>
-  )
-}
