+++ package.json
@@ -1,38 +1,29 @@
 {
-  "name": "next14correct",
+  "name": "next14ts",
   "version": "0.1.0",
-  "type": "module",
+  "private": true,
   "scripts": {
-    "sync-db": "node src/db-sync.js",
-    "seed-db": "node src/db-seed.js",
-    "unseed-db": "node src/db-unseed.js",
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "watch": "tsc -p tsconfig.json --watch"
+    "lint": "next lint"
   },
   "dependencies": {
-    "@types/react": "^18.2.48",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "i": "^0.3.7",
-    "mysql2": "^3.7.0",
+    "mysql2": "^3.7.1",
     "next": "14.0.4",
-    "nodemon": "^3.0.2",
-    "npm": "^10.2.5",
     "react": "^18",
     "react-dom": "^18",
     "sequelize": "^6.35.2"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^6.19.0",
-    "@typescript-eslint/parser": "^6.19.0",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
-    "eslint": "^8.56.0",
-    "eslint-plugin-react": "^7.33.2",
+    "eslint": "^8",
+    "eslint-config-next": "14.0.4",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
-    "typescript": "^5.3.3"
+    "typescript": "^5"
   }
 }
