--- postcss.config.tsx
@@ -1,12 +1,0 @@
-export default {
-    plugins: {
-        tailwindcss: {},
-        autoprefixer: {},
-    }
-};
-// module.exports = {
-//     plugins: {
-//         tailwindcss: {},
-//         autoprefixer: {},
-//     },
-// }
