+++ src/app/posts/[id]/page.tsx
@@ -1,4 +1,4 @@
-import {Post} from '@/app/db/post.model'
+import {Post} from '../../db/post.model'
 import Link from "next/link"
 // import {revalidatePath} from "next/cache"
