+++ src/app/posts/page.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link"
-import {Post} from "@/app/db/post.model";
+import {Post} from '../db/post.model'
 
 import {revalidatePath} from "next/cache";
