--- src/db-seed.tsx
@@ -1,3 +1,0 @@
-import {up, down} from './app/db/post.seeder.js'
-
-await up()
