--- src/db-unseed.tsx
@@ -1,3 +1,0 @@
-import {down} from './app/db/post.seeder.js'
-
-await down()
