--- tailwind.config.tsx
@@ -1,24 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-
-// module.exports = {
-export default {
-  content: [
-    './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/components/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
-  ],
-  theme: {
-    extend: {
-      backgroundImage: {
-        'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
-        'gradient-conic':
-          'conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))',
-      },
-      colors: {
-        'main-blue': '#004E98',
-        'hov-blue': '#0e3263'
-      },
-    },
-  },
-  plugins: [],
-}
